Reduce number of tests around force publishing
The `should_be_force_publishable` test macro was used in 10 controller tests.
The tests in the macro were exercising logic in the 'editions/navigation'
partial which is shared by all of these controllers. By changing these tests
to use the `GenericEditionsController` we can reduce the number of functional
tests by 54 (from 60 to 6).
