+++ test/functional/admin/case_studies_controller_test.rb
@@ -12,7 +12,6 @@ class Admin::CaseStudiesControllerTest < ActionController::TestCase
 
   should_allow_related_policies_for :case_study
   should_allow_attached_images_for :case_study
-  should_be_force_publishable :case_study
   should_prevent_modification_of_unmodifiable :case_study
   should_have_summary :case_study
   should_allow_assignment_to_document_series :case_study
