+++ test/functional/admin/consultations_controller_test.rb
@@ -22,7 +22,6 @@ class Admin::ConsultationsControllerTest < ActionController::TestCase
   should_show_inline_attachment_help_for :consultation
   should_allow_attached_images_for :consultation
   should_allow_attachment_references_for :consultation
-  should_be_force_publishable :consultation
   should_prevent_modification_of_unmodifiable :consultation
   should_allow_alternative_format_provider_for :consultation
   should_allow_assignment_to_document_series :consultation
