+++ test/functional/admin/news_articles_controller_test.rb
@@ -16,7 +16,6 @@ class Admin::NewsArticlesControllerTest < ActionController::TestCase
   should_allow_role_appointments_for :news_article
   should_allow_association_between_world_locations_and :news_article
   should_allow_attached_images_for :news_article
-  should_be_force_publishable :news_article
   should_prevent_modification_of_unmodifiable :news_article
   should_allow_overriding_of_first_published_at_for :news_article
   should_have_summary :news_article
