+++ test/functional/admin/policies_controller_test.rb
@@ -17,7 +17,6 @@ class Admin::PoliciesControllerTest < ActionController::TestCase
   should_allow_association_between_world_locations_and :policy
   should_allow_association_with_topics :policy
   should_allow_attached_images_for :policy
-  should_be_force_publishable :policy
   should_prevent_modification_of_unmodifiable :policy
   should_allow_alternative_format_provider_for :policy
   should_allow_scheduled_publication_of :policy
