+++ test/functional/admin/speeches_controller_test.rb
@@ -14,7 +14,6 @@ class Admin::SpeechesControllerTest < ActionController::TestCase
   should_allow_related_policies_for :speech
   should_allow_association_between_world_locations_and :speech
   should_allow_attached_images_for :speech
-  should_be_force_publishable :speech
   should_prevent_modification_of_unmodifiable :speech
   should_allow_overriding_of_first_published_at_for :speech
   should_allow_assignment_to_document_series :speech
